requirements.txt
# Note that these are what's required outside of the genie_python requirements
h5py
configobj
docker-compose
kafka-python
mock
parameterized
psutil
PyHamcrest
requests
six
unittest-xml-reporting
